pyproject.toml
[tool.poetry]
name = "smart-notification-service"
version = "0.1.0"
description = ""
authors = ["shashank <shashank.maurya@nxtwave.co.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
celery = {extras = ["redis"], version = "^5.5.3"}
redis = ">=4.5.5,<6.5"
django = ">=4.2"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
django-environ = "^0.12.0"
gunicorn = "^23.0.0"
requests = "^2.32.5"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
